requirements.txt
Flask==2.3.2
pdfplumber==0.7.6
spacy==3.9.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7
python-docx==0.8.11
gunicorn==21.2.0
pytest==7.4.0